+++ frontend/src/etc/helpers.ts
@@ -1,4 +1,9 @@
-import { alertFilterParams, alertRead, alertSummary } from "@/models/alert";
+import {
+  alertFilterParams,
+  alertRead,
+  alertSummary,
+  alertTreeRead,
+} from "@/models/alert";
 import { filterOption } from "@/models/base";
 import { eventFilterParams } from "@/models/event";
 import { nodeTagRead } from "@/models/nodeTag";
@@ -281,13 +286,24 @@ export function parseAlertSummary(alert: alertRead): alertSummary {
       ? alert.dispositionUser.displayName
       : "None",
     eventTime: alert.eventTime,
+    eventUuid: alert.eventUuid ? alert.eventUuid : "None",
     insertTime: alert.insertTime,
     name: alert.name,
     owner: alert.owner ? alert.owner.displayName : "None",
     queue: alert.queue.value,
     tags: alert.tags,
     tool: alert.tool ? alert.tool.value : "None",
+    toolInstance: alert.toolInstance ? alert.toolInstance.value : "None",
     type: alert.type.value,
     uuid: alert.uuid,
   };
+}
+
+export function getAllTags(alert: alertRead | alertTreeRead): nodeTagRead[] {
+  const allTags = alert.tags.concat(alert.childTags);
+
+  // Return a sorted and deduplicated list of the tags based on the tag UUID.
+  return [...new Map(allTags.map((v) => [v.uuid, v])).values()].sort((a, b) =>
+    a.value > b.value ? 1 : -1,
+  );
 }
