+++ frontend/src/services/api/alert.ts
@@ -41,38 +41,3 @@ export const Alert = {
 
   update: (data: alertUpdate[]): Promise<void> => api.update(endpoint, data),
 };
-
-export function parseAlertSummary(
-  alert: alertRead | alertTreeRead,
-): alertSummary {
-  return {
-    childTags: alert.childTags,
-    comments: alert.comments,
-    description: alert.description ? alert.description : "",
-    disposition: alert.disposition ? alert.disposition.value : "OPEN",
-    dispositionTime: alert.dispositionTime ? alert.dispositionTime : null,
-    dispositionUser: alert.dispositionUser
-      ? alert.dispositionUser.displayName
-      : "None",
-    eventTime: alert.eventTime,
-    eventUuid: alert.eventUuid ? alert.eventUuid : "None",
-    insertTime: alert.insertTime,
-    name: alert.name,
-    owner: alert.owner ? alert.owner.displayName : "None",
-    queue: alert.queue.value,
-    tags: alert.tags,
-    tool: alert.tool ? alert.tool.value : "None",
-    toolInstance: alert.toolInstance ? alert.toolInstance.value : "None",
-    type: alert.type.value,
-    uuid: alert.uuid,
-  };
-}
-
-export function getAllTags(alert: alertRead | alertTreeRead): nodeTagRead[] {
-  const allTags = alert.tags.concat(alert.childTags);
-
-  // Return a sorted and deduplicated list of the tags based on the tag UUID.
-  return [...new Map(allTags.map((v) => [v.uuid, v])).values()].sort((a, b) =>
-    a.value > b.value ? 1 : -1,
-  );
-}
